# contracts/cycle37/openapi.yaml
openapi: 3.0.3
...
info:
  title: ServiceBooking API — цикл 37 (Дома, цикл 1)
  version: "37.1.0"
  license:
    name: Proprietary
  description: |
    Конвенции проекта, действующие БЕЗ ИЗМЕНЕНИЙ:
      * camelCase; enum — строками (имена членов C#); DateTime — ISO-8601 UTC; DateOnly — "YYYY-MM-DD";
        время суток — "HH:mm";
      * ДЕНЬГИ ВЕРТИКАЛИ — ЦЕЛЫЕ РУБЛИ (integer, суффикс Rub);
      * тела осознанных 400/402/429 — ГОЛАЯ СТРОКА (text/plain) по-русски; 401/403 — пустое тело; 404 — пустое тело;
      * 451 — гейты правовых документов;
      * текст, который видит человек (статусы, причины, отказы, суммы возврата, строки суммы), собирает СЕРВЕР.

    ИСКЛЮЧЕНИЕ, КАК У ЗАКАЗОВ: все 409 маршрутов /api/stays/* — JSON с машиночитаемым code и готовым message
    (StayRefusalDto, StayGuestConflictDto, StayStaffConflictDto, StaysConflictDto).

    ГЛАВНЫЕ ПРАВИЛА, КОТОРЫЕ СХЕМА ЗАКРЕПЛЯЕТ ФОРМОЙ:
      1. Ночь D начинается в дату D; период [startDate, endDate) — endDate это дата выезда.
      2. Реквизиты для оплаты есть ТОЛЬКО в PublicStayBookingDto (по токену брони) и в PaymentDetailsDto (владельцу).
         Ни один публичный DTO дома/компании/каталога их не содержит (Т37-04).
      3. ФИО исполнителя-физлица — только в ProviderFullDto (страница брони, кабинет владельца); публично — ProviderPublicDto.
      4. Сумма возврата — «не меньше X» (StayRefundViewDto.refundAtLeastRub), считает сервер (ЮР-1).
...
servers:
  - url: http://localhost:5000
...
  - url: http://localhost:4037
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    CompanyId:
      name: companyId
      in: path
      required: true
      description: Id компании «Дома» (Company.Id, Kind = Stays). Салон/магазин/чужая → 404.
      schema: { type: string, format: uuid }
...
  responses:
    PlainTextError:
      description: Осознанный отказ — голая строка по-русски.
      content:
        text/plain:
          schema: { type: string }
...
    Unauthorized:
      description: Нет или недействителен токен. Тело пустое.
    Forbidden:
      description: Член компании без нужного права. Тело пустое.
    NotFound:
      description: Не существует или не ваше — неотличимо. Тело пустое.
...
    StaysConflict:
      description: Конфликт в кабинете.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/StaysConflictDto' }
...
    PngImage:
      description: PNG (QR-код).
      content:
        image/png:
          schema: { type: string, format: binary }
...
  schemas:
...
    StaysMyRole:
      type: string
      enum: [Owner, Manager, Housekeeper, SuperAdmin]
...
    TimeOfDay:
      type: string
      pattern: '^([01]\d|2[0-3]):[0-5]\d$'
      example: "14:00"

    # ───────────── общее ─────────────
...
    StaysConflictDto:
      type: object
      required: [code, message]
      properties:
        code: { $ref: '#/components/schemas/StaysConflictCode' }
        message: { type: string }
        conflictingPeriod: { allOf: [ { $ref: '#/components/schemas/PricePeriodDto' } ], nullable: true }
        conflicts:
          type: array
          nullable: true
          items: { $ref: '#/components/schemas/BlockBookingConflictDto' }
...
    StaysCompanyCreateInput:
      type: object
      additionalProperties: false
      required: [name, phone, ownerTermsVersion]
      properties:
        name: { type: string, minLength: 1, maxLength: 200 }
        slug: { type: string, nullable: true, maxLength: 50 }
        description: { type: string, nullable: true, maxLength: 2000 }
        phone: { type: string }
        ownerTermsVersion: { type: string }
        trialTermsVersion: { type: string, nullable: true }
    StaysTrialOutcomeDto:
      type: object
      required: [granted]
      properties:
        granted: { type: boolean }
        refusalCode: { type: string, nullable: true }
        message: { type: string, nullable: true }
        endsAtUtc: { type: string, format: date-time, nullable: true }
    StaysTrialStateDto:
      type: object
      required: [offered, eligible, durationDays]
      properties:
        offered: { type: boolean }
        eligible: { type: boolean }
        refusalCode: { type: string, nullable: true }
        message: { type: string, nullable: true }
        termsVersion: { type: string, nullable: true }
        termsText: { type: string, nullable: true }
        durationDays: { type: integer }
        endsAtUtc: { type: string, format: date-time, nullable: true }
    StaysTrialInput:
      type: object
      additionalProperties: false
      required: [termsVersion]
      properties:
        termsVersion: { type: string }
    StaysCompanyCreatedDto:
      type: object
      required: [company, token]
      properties:
        company: { $ref: '#/components/schemas/StaysCompanyManageDto' }
        token: { type: string }
        trial: { allOf: [ { $ref: '#/components/schemas/StaysTrialOutcomeDto' } ], nullable: true }
    StaysCompanyListItemDto:
      type: object
      required: [id, name, slug, publicUrl, myRole, acceptingBookings]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        slug: { type: string }
        logoUrl: { type: string, nullable: true }
        publicUrl: { type: string }
        myRole: { $ref: '#/components/schemas/StaysMyRole' }
        acceptingBookings: { type: boolean }
        awaitingPaymentCount: { type: integer, nullable: true }
    StaysSlugCheckDto:
      type: object
      required: [suggested, available]
      properties:
        suggested: { type: string }
        available: { type: boolean }
        conflict: { allOf: [ { $ref: '#/components/schemas/StaysConflictDto' } ], nullable: true }
    StaysSettingsDto:
      type: object
      required: [checkInTime, checkOutTime, minNights, maxNights, horizonDays, allowGapFill, allowSameDayCheckIn, holdMinutes, prepayPercent, cancellationPolicy, dogFeeRub, cotFeeRub, checkInInfoSendTime, checkInInfoSendFullText, arrivalReminderEnabled, housekeeperSeesGuestComment, showInCatalog]
      properties:
        checkInTime: { $ref: '#/components/schemas/TimeOfDay' }
        checkOutTime: { $ref: '#/components/schemas/TimeOfDay' }
        minNights: { type: integer, minimum: 1, maximum: 30 }
        maxNights: { type: integer, minimum: 1, maximum: 90 }
        horizonDays: { type: integer, minimum: 30, maximum: 730 }
        allowGapFill: { type: boolean }
        allowSameDayCheckIn: { type: boolean }
        holdMinutes: { type: integer, minimum: 10, maximum: 180 }
        prepayPercent: { type: integer, minimum: 0, maximum: 100 }
        cancellationPolicy: { $ref: '#/components/schemas/StayCancellationPolicy' }
        dogFeeRub: { type: integer, minimum: 0, maximum: 100000 }
        cotFeeRub: { type: integer, minimum: 0, maximum: 100000 }
        checkInInfoSendTime: { $ref: '#/components/schemas/TimeOfDay' }
        checkInInfoText: { type: string, nullable: true, maxLength: 2000 }
        checkInInfoSendFullText: { type: boolean, description: 'ЮР-4: по умолчанию false — в мессенджер уходит ссылка' }
        arrivalReminderEnabled: { type: boolean }
        housekeeperSeesGuestComment: { type: boolean, description: 'ЮР-5: по умолчанию false' }
        showInCatalog: { type: boolean }
    PaymentDetailsDto:
      type: object
      required: [paymentDetails, paymentPurpose]
      properties:
        paymentDetails: { type: string, nullable: true, maxLength: 1000 }
        paymentPurpose: { type: string, nullable: true, maxLength: 200 }
    ProviderInput:
      type: object
      additionalProperties: false
      required: [status, name, inn, claimsAddress]
      properties:
        status: { $ref: '#/components/schemas/StayProviderStatus' }
        name: { type: string, minLength: 1, maxLength: 300 }
        inn: { type: string, pattern: '^\d{10}(\d{2})?$' }
        ogrn: { type: string, nullable: true, pattern: '^\d{13}(\d{2})?$' }
        claimsAddress: { type: string, minLength: 1, maxLength: 500 }
    SlugInput:
      type: object
      additionalProperties: false
      required: [slug]
      properties:
        slug: { type: string, maxLength: 50 }
...
    StaysCompanyManageDto:
      type: object
      required: [id, name, slug, cityName, timeZoneId, isActive, showInCatalog, publicUrl, myRole, myPermissions, gate, checklist, plan]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        slug: { type: string }
        description: { type: string, nullable: true }
        phone: { type: string, nullable: true }
        email: { type: string, nullable: true }
        logoUrl: { type: string, nullable: true }
        address: { type: string, nullable: true }
        yandexMapsUrl: { type: string, nullable: true }
        twoGisUrl: { type: string, nullable: true }
        cityName: { type: string }
        timeZoneId: { type: string }
        isActive: { type: boolean }
        showInCatalog: { type: boolean }
        publicUrl: { type: string }
        myRole: { $ref: '#/components/schemas/StaysMyRole' }
        myPermissions:
          type: array
          items: { $ref: '#/components/schemas/StaysPermission' }
        settings: { allOf: [ { $ref: '#/components/schemas/StaysSettingsDto' } ], nullable: true, description: null для горничной }
        paymentDetails: { allOf: [ { $ref: '#/components/schemas/PaymentDetailsDto' } ], nullable: true, description: только при ManageCompany }
        provider: { $ref: '#/components/schemas/ProviderFullDto' }
        gate: { $ref: '#/components/schemas/GateDto' }
        checklist:
          type: array
          items: { $ref: '#/components/schemas/ChecklistItemDto' }
        plan: { $ref: '#/components/schemas/StaysPlanSummaryDto' }
        awaitingPaymentCount: { type: integer, nullable: true }
...
paths:
...
  /api/stays/companies:
    post:
      tags: [stays-company]
      operationId: staysCreateCompany
      summary: Создать компанию «Дома» (город — Шерегеш; при trialTermsVersion — попытка триала).
      security: [ { bearerAuth: [] } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/StaysCompanyCreateInput' }
      responses:
        '201':
          description: Создана.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StaysCompanyCreatedDto' }
        '400': { $ref: '#/components/responses/PlainTextError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409':
          description: Адрес занят/неверен (StaysConflictDto) или соглашение обновлено (строка).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StaysConflictDto' }
            text/plain:
              schema: { type: string }
        '503': { $ref: '#/components/responses/PlainTextError' }
  /api/stays/companies/my:
    get:
      tags: [stays-company]
      operationId: staysMyCompanies
      summary: Компании «Дома», где пользователь участник.
      security: [ { bearerAuth: [] } ]
      responses:
        '200':
          description: Список.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/StaysCompanyListItemDto' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/stays/slug-check:
    get:
      tags: [stays-company]
      operationId: staysSlugCheck
      summary: Предложить и проверить адрес компании (всегда 200).
      security: [ { bearerAuth: [] } ]
      parameters:
        - name: name
          in: query
          required: false
          schema: { type: string, maxLength: 200 }
        - name: slug
          in: query
          required: false
          schema: { type: string, maxLength: 50 }
        - name: companyId
          in: query
          required: false
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Результат.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StaysSlugCheckDto' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/stays/trial:
    get:
      tags: [stays-company]
      operationId: staysTrialState
      summary: Состояние и условия пробного периода линейки «Дома».
      security: [ { bearerAuth: [] } ]
      responses:
        '200':
          description: Состояние.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StaysTrialStateDto' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [stays-company]
      operationId: staysTrialActivate
      summary: Активировать пробный период «Домов».
      security: [ { bearerAuth: [] } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/StaysTrialInput' }
      responses:
        '200':
          description: Выдан.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StaysTrialOutcomeDto' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409':
          description: Отказ (granted false, refusalCode, message).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StaysTrialOutcomeDto' }
  /api/stays/companies/{companyId}:
    get:
      tags: [stays-company]
      operationId: staysGetCompany
      summary: Компания в кабинете (настройки, права, гейт, чек-лист, тариф).
      security: [ { bearerAuth: [] } ]
      parameters:
        - { $ref: '#/components/parameters/CompanyId' }
      responses:
        '200':
          description: Компания.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StaysCompanyManageDto' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/stays/companies/{companyId}/settings:
    put:
      tags: [stays-company]
      operationId: staysUpdateSettings
      summary: Правила компании (полная замена; действует на новые брони).
      security: [ { bearerAuth: [] } ]
      parameters:
        - { $ref: '#/components/parameters/CompanyId' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/StaysSettingsDto' }
      responses:
        '200':
          description: Компания.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StaysCompanyManageDto' }
        '400': { $ref: '#/components/responses/PlainTextError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/stays/companies/{companyId}/payment-details:
    put:
      tags: [stays-company]
      operationId: staysUpdatePaymentDetails
      summary: Реквизиты для оплаты и назначение платежа.
      security: [ { bearerAuth: [] } ]
      parameters:
        - { $ref: '#/components/parameters/CompanyId' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PaymentDetailsDto' }
      responses:
        '200':
          description: Компания.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StaysCompanyManageDto' }
        '400': { $ref: '#/components/responses/PlainTextError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/stays/companies/{companyId}/provider:
    put:
      tags: [stays-company]
      operationId: staysUpdateProvider
      summary: Сведения об исполнителе (ЮР-3).
      security: [ { bearerAuth: [] } ]
      parameters:
        - { $ref: '#/components/parameters/CompanyId' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProviderInput' }
      responses:
        '200':
          description: Компания.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StaysCompanyManageDto' }
        '400': { $ref: '#/components/responses/PlainTextError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/stays/companies/{companyId}/slug:
    put:
      tags: [stays-company]
      operationId: staysUpdateCompanySlug
      summary: Сменить адрес компании (старые ссылки и QR перестанут работать).
      security: [ { bearerAuth: [] } ]
      parameters:
        - { $ref: '#/components/parameters/CompanyId' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SlugInput' }
      responses:
        '200':
          description: Компания.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StaysCompanyManageDto' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/StaysConflict' }
  /api/stays/companies/{companyId}/qr:
    get:
      tags: [stays-company]
      operationId: staysCompanyQr
      summary: QR-код страницы компании (PNG для печати).
      security: [ { bearerAuth: [] } ]
      parameters:
        - { $ref: '#/components/parameters/CompanyId' }
      responses:
        '200': { $ref: '#/components/responses/PngImage' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/stays/companies/{companyId}/notification-settings:
    get:
      tags: [stays-company]
      operationId: staysGetNotificationSettings
      summary: Уведомления персоналу и гостям.
      security: [ { bearerAuth: [] } ]
      parameters:
        - { $ref: '#/components/parameters/CompanyId' }
      responses:
        '200':
          description: Настройки.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StaysNotificationSettingsDto' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [stays-company]
      operationId: staysUpdateNotificationSettings
      summary: Изменить уведомления (владелец).
      security: [ { bearerAuth: [] } ]
      parameters:
        - { $ref: '#/components/parameters/CompanyId' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/StaysNotificationSettingsInput' }
      responses:
        '200':
          description: Настройки.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StaysNotificationSettingsDto' }
        '400': { $ref: '#/components/responses/PlainTextError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/StaysConflict' }

  # ═════════════ кабинет: дома ═════════════
...